requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
discord.py==1.7.3
Flask==2.2.2
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gspread==5.4.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
toml==0.10.2
tomli==2.0.1
tweepy==4.10.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.2.2
yarl==1.8.1
